Fixed issue where the root node was not getting initialized.
The root node needs to be initialized now as we are starting to put
platform specific items in. To initialize the node, you must call
psapi.init_root(). To add nodes to the root, you can simply pass the node
you wish to add to the root to init_root.