Fixing tests. Doing some architectural work, going to be reworking the way the passive agent reads the config and how the config even works, in order to get it to be something better like SQlite perhaps, I'm not sure, but I would like to rethink the configuratino file being an INI file.
Either way, the config interface should be reworked. Currently its using the ConfigParser interface, which is less than ideal. Having to remember the specific method from the config parser is muddying up the logic. Having a distinct Config class that will enforce certain methods to be intact and allow an abstraction barrier will be ideal and will make the classes that use them even simpler.