Finished writing tests for the ncpa check class. Changed the return type of the returncode to unicode as by the time we actually need to calculate that value, its going to be packed into some format anyways.
Finished tests for the NRDP class as well. Stubbed out the util.send_request for the submit_to_nagios check, not really sure what a better way to test that would be. Also changed util.send_request to simply return the response as I do not believe anything was depending on that to continue working.