Updated requirements
pyOpenSSL needs to be 0.13 on all newer operating systems except CentOS 5.x and Mac OS X... to build on those systems just change the requirements.txt locally on the build box that requires 0.12 since 0.13 should always be used. The new version is REQUIRED for Debian  because it does not have OpenSSL compiled with SSLv2 - apparently a security risk. I haven't tried newer versions of it so it may be possible to upgrade to the latest.