Fixed issue with Powershell Plugins
Ended up being a completely innocuous fix. Upon inspection,
and perhaps based on powershell version, the powershell argument
was incorrect and required extensive escaping. Moving to use
the -File construct is a better way to do it and is less prone
to error.

vbs items did not suffer from this as the wscript.exe consumed the
filenames no problem.

That said, a feature was added where the user may surround $plugin_name
in the config with delimiters (quotes or double quotes) and they will
stay there after being parsed by NCPA.