Adding API Authentication Back Breaks Passive Checks
After fixing up the test runners, I found that the passive checks
were no longer running properly as they /api had recently been
fixed to request a token. Added an __INTERNAL__ variable to indicate
when an __INTERNAL__ call is going on. This requires the __INTERNAL__
variable in the listener.service module to be set to True.