Fixing Windows Logging Error
Windows was not properly logging. This was most likely
due to the recent addition of gevent getting ahold
of the global logger and redirecting its output.

Changed this by adding a new RotatingFileHandler and
setting its target at the config directed log file.

Also amended the log format to be date:level:module:message
to aid in top level debugging. This pertains to both Windows
and Linux.

Also changed both the Linux and Windows log handlers to go
to a Rotating file handlers that will rotate the logs
when the logs reach 20MB. It will keep a total of two logs
(on active, one legacy) and will overwrite the older of the two
when the younger of the two fills up.