Change test runner to be Python
Running the tests cross-platform is crucial, even though
the shell script was simple, it was never going to run on windows
without MingW or something with bash/sh. This changes it so
it can simply be run the python run_tests.py and will print
all the information we want to know.