Diskperf Being Off Breaks API Tree
On Windows, the diskperf status can bring the
API tree down and cause the listener to fail to start.
Wrapped the calls for the creating disk nodes. Decdied
to only wrap those calls as this seems to be directly
affecting disks and no other issues have arisen with CPU/Mem
etc.

If issues do arise from them, we'll create the all the nodes
in a for loop somehow and have a general way of handling such
low level errors.

Adjusted nsi so that it will automatically run diskperf -Y.
This now means that the installer will require admin privileges.