Fixing logging going to the agent/ncpa_listener.log rather than var/agent/ncpa_listener.log.
I could not completely fix this as this is caused during
service creating in the Win32Service.exe template provided
by cx_Freeze. I created a ticket for it to be fixed on the
cx_Freeze bitbucket. But thats pretty much all we can do for now.

Also added some sanity checking and logging evangelism where we
completely destroy anything that is trying to log. Also added
some PEP8 changes.