Fix Exceptions When Websocket Ends
When the api was being polled using a websocket, it would not
check for the close signal (which is None) and would simply
try to keep using it. This would cause an exception that was not
a process ender, but it did make the logs noisier and anybody
looking at them would think something was wrong.

Change the behavior to check for the None value, and close the
socket properly.