Towards debugging API plugin running changes
Currently the API is running the plugin in a spotty sense.
This commit is towards writing tests and making them a bit more
solid.

This is currently only changing the ncpa_check run_check method
to accept the config, which is necessary, as we need to pass the
config to the listener in order to set the plugin directory,
which was missing earlier.