Added SSL Directive in Config
Previously, SSL version were bound to "whatever PyOpenSSL" chose.
This obviously led to many exploits with SSLv3 and POODLE. This has
been changed so that now the user can enter the SSL version number
in the config file.

Currently supports is SSLv2, SSLv3 and TLSv1. The default is TLSv1
if no entry exists or if the entry that is entered does not associate
with an actual object recognized by the ssl module.