Adding Log max / rollover
Windows agents were properly rolling over logs but POSIX agents
were not. This was due to logmaxmb not being set in the config,
or rather no default value being set.

Also added documentation about these reasons and fixes the POSIX
agents to actually default to these values and the Windows agents
to look for non-hardcoded values.