Changing to time.time() instead of using strftime
Unfortunately it looks like strftime does not have a "%s" in Windows. Python must use the local version of the function since on Linux it looks like there is a "%s" that gives back a timestamp. Since these time differences are negligible at best, I don't see a reason why we are trying to use the datetime object instead of time. I've updated it to use time.time() to get a timestamp for the uptime portion of the API only.