Adding testing framework and documentation
Adding the ability for us to do the testing in a separate
container as to not gussy up the development and ensure that
our dev boxes aren't devving up the testing environment.

A rough draft of how it will be used is in the QUICK.md file,
while was named to be the quick documentation for the feature.