Update to use urllib2/urllib.request
Instead of using urllib's urlretrieve which seems to be causing a lot of errors while trying to connect multiple times, changed it to use urlopen which is in python 2.x in urllib2 and in python 3.x in urllib.request. Testing indicates a much better chance of successful connection than originally.