Remove psutil (installed manually) and update cffi
It looks like Python 2.7.x using cffi > 1.5.2 causes problems with dependencies. Apparently it is fixed in a newer version of python's cffi module but I am not sure. The other option is to try and get cx_Freeze to import the required parts manually. I saw an example and will try it on my test box. However, I know that using 1.5.2 allows the NCPA rpm to build and install properly.