Bugfix: ignore time on svcs pull
See https://support.nagios.com/forum/posting.php?mode=reply&f=16&t=57219

Solaris 11.4 introduces a change to the way 'svcs -a' returns data. It includes both the time and date, where Solaris 11.3 only included the time. This results in a possible extra space character in the output, resulting in NCPA returning an error on the service.

This PR should fix the issue on Oracle Solaris 11.4 by omitting the time field as it's unused. I don't have access to OpenSolaris to test the impact there, though.